Reorganization in progress.
The soiling model was organized into submodules and at "Dust" sheet was added to the QUT data to permit different properties for different experiments. Still need to check the the demo_soiling_model.ipynb notebook.

Also updated the file names in the README.md.

Changes to the optical efficiency and soiling factor computation to produce nan at night.