Added mie capabilities for fitting class.
The property loss_model of the fitting class can be either "geometry" or
"mie". In the latter case, Mie Theory will be used to compute the
effective scattering area of particles at different diameters. So far,
this only works for the fitting class and not for the full field
simulations (which must still use the geometry loss model).

The Mie capability necessitated the following major changes:

1)  The prarmeter files now have a row that specifies the loss model.
4)  The Dust tab of the experiment files must now specifty the
    refractive index real and imaginary components.
3)  The experiment files have a tab that specifies the source intensity
    distribution (which is internally normalized to integrate to one).
    This could, for example, be the reflectometer light source intensity
    distribution or the solar intensity.
4)  The extinction_weighting property specifies the area multiplier for
    each particle diameter. This must now be computed before simulation.
    In the case of the geometry loss_model, these weights are simply
    unity for all diameters.
5)  The package miepython is used for the mie computations. A reference
    will be added in subsequent commits.