Added code for paper and moved data.
Added analysis scripts for stochastic soiling paper. Moved data to a
separate repository. This data would be put into the data/ folder in
the .gitignore. The Woomera demo data has been aded to a different
folder in this repo for completeness.

Also updated the README to note that the data is stored in another repo
and needs to be placed in a data/ subfolder for the demo scripts to work.