Improved handling of extinction weight lookups.
Improved how the extinction weight lookups are
triggered. I have also added a bit of error
checking on the calling order of the sun_angles
and helios_angles methods and for when the
loss_model is not recognized and added
documentation to the compute_extinction_weights
method in the Heliostats class.