Bug fix: incidence factor moved outside of cumsum
In the computation of the prediction variance, the incidence
factor inc_ref_factor was incorrectly inside of the cumulative
sum. It is now outside the sum. This bug doesn't cause any
issue when the angle of incidence is constant, but will be
important when/if the error variance is applied to the field.